requirements.txt
# TRON — Autonomous AI Telecaller Platform
# Install with: pip install -r tron/requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
aiofiles>=23.0.0

# LiveKit (already installed via requirements.txt but listed for clarity)
livekit>=0.18.0
livekit-api>=0.7.0
